getCO2Raw returns an int, don't make it a float (#17)

+++ src/MHZ19.cpp
@@ -177,7 +177,7 @@ int MHZ19::getCO2(bool isunLimited, bool force)
     return 0;
 }
 
-float MHZ19::getCO2Raw(bool force)
+int MHZ19::getCO2Raw(bool force)
 {
     if (force == true)
         provisioning(RAWCO2);
